Final State
I'm at a point where I can start building the editor for a map. However, I want to redesign how maps are managed: instead of a grid of images, I will use a single image for the whole map. The grid will dictate where towers & monsters can be at. However, this requires me to draw complete maps, and... that's a lot, with what's going on at home currently. I may return to this project later; another project is itching for me to work on it